Added 2 new powerups and highscore system
Added two new powerups, a purple one which slows down the asteroids and a green one that destroys the asteroids in front of the ship. Added a high score system which records your highest score in a text file and also contains what powerups you used on that run. Aslo streamlined some of the powerup code and removed some unneeded lines.